Move bulk_invoke's implementation into a header underneath bulk_functions/
